refactor: reorganizes the project structure to improve organization and maintenance
