+++ src/assets/styles/global.css
@@ -1,6 +1,7 @@
 .red-line {
   width: 80px;
   height: 4px;
+  border-radius: 20%;
   background-color: #FF0000;
   display: block;
 }
@@ -17,5 +18,5 @@
   height: 10px;
   border-radius: 50%;
   background-color: #FF0000;
-  margin-right: 10px;
+  margin-right: 15px;
 }
